Revert "refactor(imports): allow importing from hooks dir"
This reverts commit ba876c449d3dce79972f53ca23dc9866e7b7ab84.
