+++ src/components/FormFields/LabeledOptionsField.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { mount } from 'enzyme'
 import { ReactFinalForm } from '@dhis2/ui'
-import { useParameterOption } from '../../hooks'
+import { useParameterOption } from '../../hooks/parameter-options'
 import LabeledOptionsField from './LabeledOptionsField'
 
 const { Form } = ReactFinalForm
