+++ src/components/FormFields/ScheduleField.js
@@ -2,7 +2,7 @@ import React from 'react'
 import i18n from '@dhis2/d2-i18n'
 import { NoticeBox } from '@dhis2/ui'
 import { PropTypes } from '@dhis2/prop-types'
-import { useJobType } from '../../hooks'
+import { useJobType } from '../../hooks/job-types'
 import CronField from './CronField'
 import DelayField from './DelayField'
